+++ lib/src/web/readers/messages/add_global_contexts_message_reader.dart
@@ -9,15 +9,16 @@
 // "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the Apache License Version 2.0 for the specific language governing permissions and limitations there under.
 
-import Foundation
-import SnowplowTracker
+import 'package:flutter/foundation.dart';
 
-struct CreateTrackerMessageReader: Decodable {
-    let namespace: String
-    let networkConfig: NetworkConfigurationReader
-    let trackerConfig: TrackerConfigurationReader?
-    let subjectConfig: SubjectConfigurationReader?
-    let gdprConfig: GdprConfigurationReader?
-    let emitterConfig: EmitterConfigurationReader?
-    let globalContextsConfig: GlobalContextsConfigurationReader?
+@immutable
+class AddGlobalContextsMessageReader {
+  final String tracker;
+  final String tag;
+  final dynamic context;
+
+  AddGlobalContextsMessageReader(dynamic map)
+      : tracker = map['tracker'],
+        tag = map['tag'],
+        context = map['context'];
 }